Added Oracle Added Maven

+++ src/atm/Card.java
@@ -9,16 +9,16 @@ import java.util.Random;
  */
 public class Card {
     
-    private static short cardNum;
+    private static int cardNum;
     
-    public short card(){
+    public int card(){
         
         /*
         This class randomly generates a 4 digit card number to apply to each
         new user.
         */
         Random r = new Random();
-        cardNum = (short)r.nextInt(9999);
+        cardNum = r.nextInt(9999);
         System.out.println("Your card number is " + cardNum + "\n");
         return cardNum;
     }
